iraceplot: Plots for Visualizing the Data Produced by the Irace Package

Graphical visualization tools for analysing the data produced by irace.

Graphical Visualization Tools for Analysing the Data Produced by Irace.

boxplot_test; scatter_test; parallel_coord; parallel_coord2; parallel_cat; sampling_frequency; sampling_frequency_iteration; sampling_pie; plot_experiments_matrix; sampling_distance; sampling_heatmap; sampling_heatmap2; boxplot_training; scatter_training; boxplot_performance; scatter_performance; report; plot_model;

If you need information about any function you can write: ?name_function

If you need more information, go to the following page: https://auto-optimization.github.io/iraceplot/

Details

License: MIT + file LICENSE

See also

Author

Maintainer: Leslie Pérez Cáceres leslie.perez@pucv.cl

Authors:

Pablo Oñate Marín pablo.onate.m@gmail.com

Manuel López-Ibáñez manuel.lopez-ibanez@uma.es (ORCID)

Maintainers: Pablo Oñate Marín and Leslie Pérez Cáceres leslie.perez@gmail.com