Plot parameter configurations using parallel coordinates

Parallel coordinates plot of a set of provided configurations. Each line in the plot represents a configuration. The parameters to be included in the plot can be selected with the param_names argument. Additionally, the maximum number of parameters to be displayed in one plot. A list of plots is returned by this function in several plots are required to display the selected data.

plot_configurations(
  configurations,
  parameters,
  param_names = parameters$names,
  by_n_param = NULL,
  filename = NULL
)

Arguments

configurations

Data frame, configurations in irace format (example: configurations = iraceResults$allConfigurations)

parameters

(list()) Parameter space in irace format. See the function irace::readParameters().

param_names

(character()) Parameters to be included in the plot. Example: c("algorithm","alpha","rho","q0","rasrank").

by_n_param

Numeric (optional), maximum number of parameters to be displayed

filename

(character(1)) File name to save the plot, for example "~/path/example/filename.png".

Value

parallel coordinates plot

Details

To export the plot to a file, it is possible to do it so manually using the functionality provided by plotly in the plot. If a filename is provided, orca server will be used to export the plots and thus, it requires the library to be installed (https://github.com/plotly/orca).

Examples

iraceResults <- read_logfile(system.file(package="irace", "exdata",
                                         "irace-acotsp.Rdata", mustWork = TRUE))
parameters <- iraceResults$scenario$parameters
plot_configurations(iraceResults$allConfigurations[iraceResults$iterationElites,], 
                parameters = parameters)

plot_configurations(iraceResults$allConfigurations[iraceResults$iterationElites,], 
                parameters = parameters,
                param_names = c("algorithm", "alpha", "rho", "q0", "rasrank"))

plot_configurations(iraceResults$allConfigurations[iraceResults$iterationElites,], 
                parameters = parameters, by_n_param = 5)
#> [[1]]
#> 
#> [[2]]
#> 
#> [[3]]
#>