Print parameter space in the textual format accepted by irace.

Print parameter space in the textual format accepted by irace.

printParameters(params)

Arguments

params

(list()) Parameter object stored in irace.Rdata or read with irace::readParameters().

Examples

load(system.file(package="iraceplot", "exdata", "guide-example.Rdata", mustWork = TRUE))
printParameters(iraceResults$parameters)
#>   algorithm             "--" c (as,mmas,eas,ras,acs)
#> localsearch "--localsearch " c       (0,1,2,3)
#>       alpha       "--alpha " r           (0,5)
#>        beta        "--beta " r          (0,10)
#>         rho        "--rho  " r        (0.01,1)
#>        ants        "--ants " i         (5,100)
#>        nnls        "--nnls " i          (5,50) | localsearch %in% c(1, 2, 3)
#>          q0          "--q0 " r           (0,1) | algorithm %in% c("acs")
#>         dlb         "--dlb " c           (0,1) | localsearch %in% c(1, 2, 3)
#>     rasrank    "--rasranks " i         (1,100) | algorithm %in% c("ras")
#> elitistants "--elitistants " i         (1,750) | algorithm %in% c("eas")