Create HTML Report from irace data

This function creates an HTML report of the most relevant irace data. This report provides general statistics and plots that show the best configurations and their performance. Example: https://auto-optimization.github.io/iraceplot/articles/example/report_example.html

report(
  irace_results,
  filename = "report",
  sections = list(experiments_matrix = NULL),
  interactive = base::interactive()
)

Arguments

irace_results

The data generated when loading the Rdata file created by irace

filename

(character(1)) Filename indicating where to save the report (example: "~/path-to/filename")

sections

(list()) List of sections to enable/disable. This is useful for disabling sections that may cause problems, such as out-of-memory errors. NA means automatically enable/disable a section depending on the memory required.

interactive

(logical(1)) TRUE if the plot is generated interactively (using plotly::ggplotly() or plotly::plot_ly()) or FALSE it is generated statically. Defaults to the value returned by interactive(),

Value

filename where the report was created or it opens the report in the default browser (interactive)

Examples

# \dontrun{
load(system.file(package="irace", "exdata", "irace-acotsp.Rdata", mustWork = TRUE))
report(iraceResults, filename="report")
#> Creating file/tmp/RtmpUzHWbF/file35077f594f5/reference/report.html
#> 
#> 
#> processing file: report_html.Rmd
#>   |                                                                              |                                                                      |   0%  |                                                                              |..                                                                    |   3%
#>   ordinary text without R code
#> 
#>   |                                                                              |....                                                                  |   5%
#> label: unnamed-chunk-1 (with options) 
#> List of 2
#>  $ echo  : logi FALSE
#>  $ engine: chr "css"
#> 
#>   |                                                                              |.....                                                                 |   8%
#>   ordinary text without R code
#> 
#>   |                                                                              |.......                                                               |  10%
#> label: unnamed-chunk-2 (with options) 
#> List of 1
#>  $ include: logi FALSE
#> 
#>   |                                                                              |.........                                                             |  13%
#>   ordinary text without R code
#> 
#>   |                                                                              |...........                                                           |  15%
#> label: setup (with options) 
#> List of 1
#>  $ include: logi FALSE
#> 
#>   |                                                                              |.............                                                         |  18%
#>   ordinary text without R code
#> 
#>   |                                                                              |..............                                                        |  21%
#> label: unnamed-chunk-3 (with options) 
#> List of 3
#>  $ echo: logi TRUE
#>  $ eval: logi FALSE
#>  $ code: chr [1:60] "## irace scenario:" "scenarioFile = \"./scenario.txt\"" "parameterFile = \"./parameters.txt\"" "execDir = \"./\"" ...
#> 
#>   |                                                                              |................                                                      |  23%
#>   ordinary text without R code
#> 
#>   |                                                                              |..................                                                    |  26%
#> label: printParameters (with options) 
#> List of 2
#>  $ echo   : logi FALSE
#>  $ results: chr "asis"
#> 
#>   |                                                                              |....................                                                  |  28%
#>    inline R code fragments
#> 
#>   |                                                                              |......................                                                |  31%
#> label: unnamed-chunk-4 (with options) 
#> List of 2
#>  $ fig.align: chr "center"
#>  $ echo     : logi FALSE
#> 
#> Quitting from lines 156-158 (report_html.Rmd) 
#> Error in knitr::kable(x, row.names = row.names): argument "row.names" is missing, with no default
#> 
# }