Sampling pie plot

This function creates a pie plot of the values sampled of a set of selected parameters. Numerical parameters are discretized to maximum n_bins intervals. The size of the slices are proportional to the number of configurations that have assigned a parameter value within the rank or the value assigned to that slice. Parameters can be selected by providing their names in the param_names argument.

sampling_pie(irace_results, param_names = NULL, n_bins = 3, filename = NULL)

Arguments

irace_results

The data generated when loading the .Rdata file created by irace (or the filename of that file).

param_names

String vector, A set of parameters to be included (example: param_names = c("algorithm","dlb"))

n_bins

Numeric (default 3), number of intervals to generate for numerical parameters.

filename

(character(1)) File name to save the plot, for example "~/path/example/filename.png".

Value

Sampling pie plot

Examples

iraceResults <- read_logfile(system.file(package="irace", "exdata",
                                         "irace-acotsp.Rdata", mustWork = TRUE))
sampling_pie(iraceResults)

# \donttest{
sampling_pie(iraceResults, param_names = c("algorithm", "dlb", "ants"))

# }