Parallel Coordinates Plot

The `parallel_coord` function creates a parallel coordinates plot of a set of selected configurations. Each line in the plot represents a configuration. By default, the final elite configurations are shown. To visualize configurations of other iterations these must be provided setting the argument iterations, configurations of different iterations are shown in different colors. Setting the only_elites argument to FALSE allows to display all configurations in the selected iterations, specific configurations can be selected providing their ids in the id_configuration argument.

The parameters to be included in the plot can be selected with the param_names argument. Additionally, the maximum number of parameters to be displayed in one plot. A list of plots is returned by this function in several plots are required to display the selected data.

To export the plot to a file, it is possible to do it so manually using the functionality provided by plotly in the plot. If a filename is provided, orca server will be used to export the plots and thus, it requires the library to be installed (https://github.com/plotly/orca).

parallel_coord(
  irace_results,
  id_configuration = NULL,
  param_names = NULL,
  iterations = NULL,
  only_elite = TRUE,
  by_n_param = NULL,
  color_by_instances = TRUE,
  filename = NULL
)

Arguments

irace_results

The data generated when loading the Rdata file created by irace

id_configuration

Numeric vector, configurations ids whose performance should be included in the plot (example: id_configuration = c(20,50,100,300,500,600,700))

param_names

String vector, names of the parameters that should be included in the plot (example: param_names = c("algorithm","alpha","rho","q0","rasrank"))

iterations

Numeric vector, iteration number that should be included in the plot (example: iterations = c(1,4,5))

only_elite

logical (default TRUE), only print elite configurations (argument ignored when id_configuration is provided)

by_n_param

Numeric (optional), maximum number of parameters to be displayed.

color_by_instances

Logical (default TRUE), choose how to color the lines. TRUE shows the number of instances evaluated by the configuration in the colores. FALSE to show the iteration number where the configuration was sampled.

filename

(character(1)) File name to save the plot, for example "~/patch/example/filename.png".

Value

parallel coordinates plot

Examples

parallel_coord(iraceResults)

if (FALSE) { 
parallel_coord(iraceResults, by_n_param = 5)
parallel_coord(iraceResults, only_elite = FALSE)
parallel_coord(iraceResults, id_configuration = c(20, 50, 100, 300, 500, 600, 700))
parallel_coord(iraceResults, param_names = c("algorithm", "alpha", "rho", "q0", "rasrank"))
parallel_coord(iraceResults, iterations = c(1, 4, 6))
}